data/v6/requirements.txt
datasets
requests
tqdm
ollama
Pillow
